Fix missing last page when using multipage
In the specific case where fetching more pages is stopped due to finding
a duplicate link, the link to the page where the duplicate found was not
being returned.

This manifested as the last page of a multipage article not being
processed.

This commit fixes this by returning an array with the current link in it
when finding a duplicate link.